started symbol table entry
